agent/pyproject.toml
[project]
name = "agent"
version = "0.1.5"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.16.0",
    "pandas>=2.3.3",
    "pandas-gbq>=0.29.2",
    "pyarrow>=21.0.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "folium>=0.14.0",
    "streamlit-folium>=0.15.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "streamlit-chat>=0.1.1",
    "streamlit-option-menu>=0.3.6",
    "streamlit-aggrid>=0.3.4",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "data_agent",
    "health_agent",
    "insights_agent"
]
